Add files via upload
Symmetric Key Encryption and Decryption with Fernet

Introduction:
Symmetric key encryption plays a pivotal role in securing sensitive information by using the same key for both encryption and decryption processes. In this report, we explore the implementation of symmetric key encryption and decryption using the Fernet encryption method, a part of the Python cryptography library.